ocr_markdown_pipeline/pyproject.toml
[tool.poetry]
name = "ocr_markdown_pipeline"
version = "0.1.0"
description = "OCR-to-Markdown Pipeline using crewAI"
authors = ["Jarad DeLorenzo <jaradd@gmail.com>"]
packages = [
    { include = "ocr_markdown_pipeline", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = ">=0.79.2,<1.0.0" }
asyncio = "*"
pytesseract = "*"
Pillow = "*"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
ocr_markdown_pipeline = "ocr_markdown_pipeline.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]
